Move template to s2s-utils
