+++ packages/babel-plugin-s2s-action-types/package.json
@@ -21,7 +21,7 @@
   "dependencies": {
     "babel-add-flow-comments": "^0.2.0",
     "babel-plugin-syntax-flow": "^6.18.0",
-    "babel-template": "^6.26.0",
+    "s2s-utils": "^0.0.0",
     "babel-types": "^6.26.0",
     "lodash.snakecase": "^4.1.1",
     "s2s-babel-flow-types": "^0.1.0"
