+++ packages/babel-plugin-s2s-reducer-root/package.json
@@ -21,7 +21,6 @@
   "dependencies": {
     "babel-add-flow-comments": "^0.2.0",
     "babel-plugin-syntax-flow": "^6.18.0",
-    "babel-template": "^6.26.0",
     "babel-types": "^6.26.0",
     "globby": "^6.1.0",
     "s2s-babel-flow-types": "^0.1.0",
